# Conflicts:
#	test/UniswapV3Pair.spec.ts
#	test/__snapshots__/UniswapV3Factory.spec.ts.snap
#	test/__snapshots__/UniswapV3Pair.spec.ts.snap
#	test/shared/utilities.ts

+++ contracts/interfaces/IUniswapV3Pair.sol
@@ -3,7 +3,7 @@ pragma solidity >=0.5.0;
 pragma experimental ABIEncoderV2;
 
 interface IUniswapV3Pair {
-    event Initialized(uint128 sqrtPrice);
+    event Initialized(uint160 sqrtPrice);
 
     // event PositionSet(address owner, int24 tickLower, int24 tickUpper, uint8 feeVote, int112 liquidityDelta);
 
@@ -33,7 +33,7 @@ interface IUniswapV3Pair {
 
     function liquidityCurrent() external view returns (uint128);
 
-    function sqrtPriceCurrent() external view returns (uint128);
+    function sqrtPriceCurrent() external view returns (uint160);
 
     function feeGrowthGlobal0() external view returns (uint256);
 
@@ -51,7 +51,7 @@ interface IUniswapV3Pair {
     function getCumulatives() external view returns (uint32 blockTimestamp, int56 tickCumulative);
 
     // initialize the pair
-    function initialize(uint128 sqrtPrice) external;
+    function initialize(uint160 sqrtPrice) external;
 
     // collect fees
     function collectFees(
